// projects/ngx-grid-core/src/lib/ui/themes/ngx-grid-theme.interface.ts
export interface INgxGridTheme {
    appBackground: string
    textColor: string
    fontSize: string
    borderColor: string
    gridFocusedShadow: string
    gridBorderTopWidth: string
    gridBorderRightWidth: string
    gridBorderBottomWidth: string
    gridBorderLeftWidth: string
    rowHeight: string
    rowBackgroundColorFocused: string
    rowBackgroundColorSelected: string
    rowBackgroundColorPreSelected: string
    rowBackgroundColorHover: string
    floatingTitleBackgroundColorHover: string
    rowThumbWidth: string
    rowThumbBaseBackgroundColor: string
    rowThumbBackgroundColor: string
    rowThumbBackgroundColorHover: string
    rowThumbBackgroundColorSelected: string
    rowThumbBorderColorSelected: string
    rowThumbBorderSelectedWidth: string
    rowThumbActionTextColor: string
    rowThumbActionTextColorHover: string
    rowThumbBorderRightWidth: string
    rowThumbBorderBottomWidth: string
    columnBackgroundColorSelected: string
    columnBorderColorSelected: string
    columnBorderSelectedWidth: string
    cellCursor: string
    cellBackgroundColor: string
    cellOutlineColorSelectedHover: string
    cellOutlineWidth: string
    cellBorderColor: string
    cellBorderHorizontalWidth: string
    cellBorderVerticalWidth: string
    cellBackgroundColorEditable: string
    cellButtonBackgroundColor: string
    cellButtonBackgroundColorHover: string
    chipInputBackdropColor: string
    fileDropzoneBackdropColor: string
    fileDropzoneBackgroundColor: string
    fileDropzoneTextColor: string
    selectionInset: string
    selectionOverlappingBorderColor: string
    selectionOverlappingBackgroundColorSelected: string
    selectionOverlappingBackgroundColorFocusedSelected: string
    selectionBorderWidth: string
    selectionBorderTopStyle: string
    selectionBorderRightStyle: string
    selectionBorderLeftStyle: string
    selectionBorderBottomStyle: string
    selectionMultiselectBorderColor: string
    selectionMultiselectBorderColorFocused: string
    selectionMultiselectBorderWidth: string
    cellActionButtonBackground: string
    cellActionButtonBackgroundHover: string
    contextMenuItemBackgroundColor: string
    headerColumnFontSize: string
    headerColumnFontWeight: string
}